+++ src/app/app.component.html
@@ -1,6 +1,4 @@
 <app-header></app-header>
-<adsk-forge-viewer [viewerOptions]="viewerOptions"
-                   (onSelectionChanged)="selectionChanged($event)"></adsk-forge-viewer>
 <app-main></app-main>>
