+++ src/app/shared/components/main/main.component.css
@@ -1,3 +1,11 @@
+
+#mainForgeViewerDiv{
+  position: relative;
+  height: 100%;
+  
+}
+
+
 .container-md{
   justify-self: center;
   justify-content: center;
@@ -164,6 +172,7 @@ input[type=search] {
   top: 10%;
   left: 10%;
   width: 80%;
+  
   max-height: 80%;
   z-index: 10;
   background-color: #fff;
@@ -171,7 +180,10 @@ input[type=search] {
 
 .classViewGrid{
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(150px, 50%));
+  grid-template-columns: 70% 30%;
+  height: 80%;
+  /*
+  grid-template-columns: repeat(auto-fill, minmax(150px, 50%));*/
 }
 
 .imgcontainer{
