+++ src/app/shared/components/main/main.component.spec.ts
@@ -1,5 +1,5 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
 
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MainComponent } from './main.component';
 
 describe('MainComponent', () => {
