+++ src/index.html
@@ -9,6 +9,5 @@
 </head>
 <body [style.overflow]="bodyOverflow">
   <app-root></app-root>
-  <my-app>loading</my-app>
 </body>
 </html>
